refactor(automation): extract ImplementationPhaseRunner
Moves the per-issue 6-phase pipeline (plan -> impl -> review-loop ->
test -> PR -> follow-up -> learn) and 20+ phase helpers from
``IssueImplementer`` into a new ``ImplementationPhaseRunner`` class in
``implementer_phase_runner.py``. The runner holds a back-reference to
the parent ``IssueImplementer`` so cross-method dispatch flows back
through the coordinator's thin shims -- this preserves the test-patch
contract: ``patch.object(impl, "_has_plan", ...)`` continues to
intercept every call site, including the ones inside ``_implement_issue``
that were moved here.

Patchable module-level symbols (``is_plan_review_approved``,
``fetch_issue_info``, ``invoke_claude_with_session``, ``get_repo_slug``,
``current_trunk_githash``, ``AGENT_IMPLEMENTER``) are re-exported by
``implementer.py`` and looked up via the runner's ``_impl_module``
helper, so existing ``patch("hephaestus.automation.implementer.X")``
test idioms keep working unchanged.

``test_phase_agent_wiring.py`` is updated to inspect both
``implementer.py`` and its newly-extracted companion
``implementer_phase_runner.py`` when validating that the implementer
phase passes ``agent=AGENT_IMPLEMENTER`` to ``invoke_claude_with_session``.

implementer.py shrinks from 1974 LOC to 883 LOC (a 55% reduction);
the remaining body is the coordinator surface (``__init__``, ``run``,
``_load_issues``, ``_health_check``, ``_analyze_dependencies``,
``_implement_all``), thin delegations to the three collaborators, and
the CLI machinery (``_setup_logging``, ``_parse_args``, ``main``).

Part of #597.
