+++ tests/unit/automation/test_phase_agent_wiring.py
@@ -31,10 +31,20 @@ AUTOMATION_DIR = Path(automation_pkg.__file__).parent
 
 
 # Self-agent phases: module owns a unique session identity.
-SELF_AGENT_PHASES: list[tuple[str, str]] = [
-    ("plan_reviewer.py", "AGENT_PLAN_REVIEWER"),
-    ("pr_reviewer.py", "AGENT_PR_REVIEWER"),
-    ("implementer.py", "AGENT_IMPLEMENTER"),
+#
+# Each entry is ``(module_file, expected_agent_constant, companion_files)``,
+# where ``companion_files`` is an optional tuple of sibling modules that
+# share the same self-agent identity. For ``implementer.py`` the
+# per-issue phase runner was extracted into ``implementer_phase_runner.py``
+# in #597 — both files participate in the implementer's session and must
+# be inspected together. ``invoke_claude_with_session`` is called inside
+# the runner; the ``AGENT_IMPLEMENTER`` constant is referenced through
+# the implementer module's namespace (``_impl_mod.AGENT_IMPLEMENTER``)
+# so test patches at ``implementer.AGENT_IMPLEMENTER`` still take effect.
+SELF_AGENT_PHASES: list[tuple[str, str, tuple[str, ...]]] = [
+    ("plan_reviewer.py", "AGENT_PLAN_REVIEWER", ()),
+    ("pr_reviewer.py", "AGENT_PR_REVIEWER", ()),
+    ("implementer.py", "AGENT_IMPLEMENTER", ("implementer_phase_runner.py",)),
 ]
 
 
@@ -46,36 +56,63 @@ CONTINUATION_PHASES: list[str] = [
 ]
 
 
-@pytest.mark.parametrize("module_file, expected_agent", SELF_AGENT_PHASES)
-def test_self_agent_phase_imports_expected_agent(module_file: str, expected_agent: str) -> None:
-    """Each self-agent phase imports its dedicated AGENT_* constant."""
-    src = (AUTOMATION_DIR / module_file).read_text()
+def _read_phase_sources(module_file: str, companions: tuple[str, ...]) -> str:
+    """Return the concatenated source of *module_file* and any companions."""
+    parts = [(AUTOMATION_DIR / module_file).read_text()]
+    parts.extend((AUTOMATION_DIR / c).read_text() for c in companions)
+    return "\n".join(parts)
+
+
+@pytest.mark.parametrize("module_file, expected_agent, companions", SELF_AGENT_PHASES)
+def test_self_agent_phase_imports_expected_agent(
+    module_file: str, expected_agent: str, companions: tuple[str, ...]
+) -> None:
+    """Each self-agent phase imports its dedicated AGENT_* constant.
+
+    Imports may live in ``module_file`` itself or in any of its
+    ``companions`` (e.g. ``implementer_phase_runner.py`` was extracted from
+    ``implementer.py`` in #597 but still participates in the implementer's
+    session identity).
+    """
+    src = _read_phase_sources(module_file, companions)
     import_pattern = re.compile(rf"from\s+\.session_naming\s+import\s+[^\n]*\b{expected_agent}\b")
     assert import_pattern.search(src), (
-        f"{module_file} must import {expected_agent} from .session_naming"
+        f"{module_file} (and companions {companions}) must import {expected_agent} "
+        f"from .session_naming"
     )
 
 
-@pytest.mark.parametrize("module_file, expected_agent", SELF_AGENT_PHASES)
+@pytest.mark.parametrize("module_file, expected_agent, companions", SELF_AGENT_PHASES)
 def test_self_agent_phase_passes_expected_agent_kwarg(
-    module_file: str, expected_agent: str
+    module_file: str, expected_agent: str, companions: tuple[str, ...]
 ) -> None:
-    """Each self-agent phase passes its AGENT_* constant via ``agent=``."""
-    src = (AUTOMATION_DIR / module_file).read_text()
-    kwarg_pattern = re.compile(rf"\bagent\s*=\s*{expected_agent}\b")
+    """Each self-agent phase passes its AGENT_* constant via ``agent=``.
+
+    The implementer module dispatches the actual call through
+    ``implementer_phase_runner`` and references the constant as
+    ``_impl_mod.AGENT_IMPLEMENTER`` so test patches at
+    ``hephaestus.automation.implementer.AGENT_IMPLEMENTER`` continue to
+    work. The pattern below accepts both the bare ``AGENT_IMPLEMENTER``
+    form and the namespaced ``X.AGENT_IMPLEMENTER`` form.
+    """
+    src = _read_phase_sources(module_file, companions)
+    kwarg_pattern = re.compile(rf"\bagent\s*=\s*(?:[A-Za-z_][A-Za-z0-9_]*\.)?{expected_agent}\b")
     assert kwarg_pattern.search(src), (
-        f"{module_file} must pass agent={expected_agent} to invoke_claude_with_session"
+        f"{module_file} (and companions {companions}) must pass agent={expected_agent} "
+        "to invoke_claude_with_session"
     )
 
 
-@pytest.mark.parametrize("module_file, expected_agent", SELF_AGENT_PHASES)
-def test_self_agent_phase_does_not_use_foreign_agent(module_file: str, expected_agent: str) -> None:
+@pytest.mark.parametrize("module_file, expected_agent, companions", SELF_AGENT_PHASES)
+def test_self_agent_phase_does_not_use_foreign_agent(
+    module_file: str, expected_agent: str, companions: tuple[str, ...]
+) -> None:
     """A self-agent phase must not pass any other AGENT_* constant."""
-    src = (AUTOMATION_DIR / module_file).read_text()
-    found = set(re.findall(r"\bagent\s*=\s*(AGENT_[A-Z_]+)\b", src))
+    src = _read_phase_sources(module_file, companions)
+    found = set(re.findall(r"\bagent\s*=\s*(?:[A-Za-z_][A-Za-z0-9_]*\.)?(AGENT_[A-Z_]+)\b", src))
     assert found <= {expected_agent}, (
-        f"{module_file} uses unexpected AGENT_* constants: "
-        f"{found - {expected_agent}}; expected only {expected_agent}"
+        f"{module_file} (and companions {companions}) uses unexpected AGENT_* "
+        f"constants: {found - {expected_agent}}; expected only {expected_agent}"
     )
